🚀 Fix: suspense 적용을 위한 서버&클라이언트 컴포넌트 분리#22
